# Padding for IO buffers.