--- a/dbms/src/IO/BufferWithOwnMemory.h
+++ b/dbms/src/IO/BufferWithOwnMemory.h
@@ -26,6 +26,7 @@ namespace DB
   */
 struct Memory : boost::noncopyable, Allocator<false>
 {
+    static constexpr size_t pad_right = 15;
     size_t m_capacity = 0;
     size_t m_size = 0;
     char * m_data = nullptr;
@@ -72,17 +73,17 @@ struct Memory : boost::noncopyable, Allocator<false>
             m_size = m_capacity = new_size;
             alloc();
         }
-        else if (new_size <= m_capacity)
+        else if (new_size <= m_size)
         {
             m_size = new_size;
             return;
         }
         else
         {
-            new_size = align(new_size, alignment);
-            m_data = static_cast<char *>(Allocator::realloc(m_data, m_capacity, new_size, alignment));
-            m_capacity = new_size;
-            m_size = m_capacity;
+            size_t new_capacity = align(new_size + pad_right, alignment);
+            m_data = static_cast<char *>(Allocator::realloc(m_data, m_capacity, new_capacity, alignment));
+            m_capacity = new_capacity;
+            m_size = m_capacity - pad_right;
         }
     }
 
@@ -103,13 +104,15 @@ private:
             return;
         }
 
-        ProfileEvents::increment(ProfileEvents::IOBufferAllocs);
-        ProfileEvents::increment(ProfileEvents::IOBufferAllocBytes, m_capacity);
+        size_t padded_capacity = m_capacity + pad_right;
 
-        size_t new_capacity = align(m_capacity, alignment);
+        ProfileEvents::increment(ProfileEvents::IOBufferAllocs);
+        ProfileEvents::increment(ProfileEvents::IOBufferAllocBytes, padded_capacity);
+
+        size_t new_capacity = align(padded_capacity, alignment);
         m_data = static_cast<char *>(Allocator::alloc(new_capacity, alignment));
         m_capacity = new_capacity;
-        m_size = m_capacity;
+        m_size = m_capacity - pad_right;
     }
 
     void dealloc()
@@ -137,6 +140,7 @@ public:
         : Base(nullptr, 0), memory(existing_memory ? 0 : size, alignment)
     {
         Base::set(existing_memory ? existing_memory : memory.data(), size);
+        Base::padded = !existing_memory;
     }
 };