@alexey-milovidov 2018-08-30
Ok.

But the code in `Memory` is a little bit tangled (from my point of view).
Because capacity is first initialized to value without padding and then padded.